// src/types/requestLesson.dto.ts
export interface RequstLessonDto {
  title: string; // Тема занятия. Одинаковая на все создаваемые занятия
  days: number[]; // Дни недели, по которым нужно создать занятия, где 0 - это воскресенье
  lastDate?: string; // Последняя дата, до которой нужно создавать занятия.
  firstDate: string; // Первая дата, от которой нужно создавать занятия
  teacherIds: number[]; // id учителей, ведущих занятия
  lessonsCount?: number; // Количество занятий для создания
}